requirements.txt
# Зависимости для Chamby Faucet Bot

# Telegram Bot API
python-telegram-bot==13.7

# HTTP запросы
requests==2.26.0

# SQLite3 встроен в Python 3.8, дополнительная установка не требуется

# Опционально: для реальной работы с TON блокчейном
# Раскомментируйте нужную библиотеку после настройки:

# tonsdk==1.0.15
# или
# pytonlib==0.0.41
# или
# tonpy==0.1.0

# Для работы с переменными окружения (рекомендуется)
python-dotenv==0.19.0
